examples/pipelines/langgraph/requirements.txt
fastapi
langgraph-agent
langchain-core
langgraph
langchain-anthropic
pydantic
python-dotenv
requests
typing-extensions